Move exported methods to utils + validators and their tests
