+++ datadog/internal/utils/utils_test.go
@@ -1,0 +1,37 @@
+package utils
+
+import (
+	"fmt"
+	"testing"
+)
+
+func TestAccountAndLambdaArnFromID(t *testing.T) {
+	cases := map[string]struct {
+		id        string
+		accountId string
+		lambdaArn string
+		err       error
+	}{
+		"basic":               {"123456789 /aws/lambda/my-fun-funct", "123456789", "/aws/lambda/my-fun-funct", nil},
+		"no delimiter":        {"123456789", "", "", fmt.Errorf("error extracting account ID and Lambda ARN from an AWS integration id: 123456789")},
+		"multiple delimiters": {"123456789 extra bits", "", "", fmt.Errorf("error extracting account ID and Lambda ARN from an AWS integration id: 123456789 extra bits")},
+	}
+	for name, tc := range cases {
+		accountId, lambdaArn, err := AccountAndLambdaArnFromID(tc.id)
+
+		if err != nil && tc.err != nil && err.Error() != tc.err.Error() {
+			t.Errorf("%s: errors should be '%s', not `%s`", name, tc.err.Error(), err.Error())
+		} else if err != nil && tc.err == nil {
+			t.Errorf("%s: errors should be nil, not `%s`", name, err.Error())
+		} else if err == nil && tc.err != nil {
+			t.Errorf("%s: errors should be '%s', not nil", name, tc.err.Error())
+		}
+
+		if accountId != tc.accountId {
+			t.Errorf("%s: account ID '%s' didn't match `%s`", name, accountId, tc.accountId)
+		}
+		if lambdaArn != tc.lambdaArn {
+			t.Errorf("%s: lambda arn '%s' didn't match `%s`", name, lambdaArn, tc.lambdaArn)
+		}
+	}
+}
