+++ datadog/tests/resource_datadog_role_test.go
@@ -95,7 +95,7 @@ func testCheckRolePermission(rolename string, permissionsSource string, permissi
 		perm := map[string]interface{}{
 			"id": permissionID,
 		}
-		permissionIDHash := schema.HashResource(datadog.GetRolePermissionSchema())(perm)
+		permissionIDHash := schema.HashResource(utils.GetRolePermissionSchema())(perm)
 
 		return resource.TestCheckResourceAttr(rolename, fmt.Sprintf("permission.%d.id", permissionIDHash), permissionID)(s)
 	}
